DRY applied
